refactor: create exceptions and content_types
