+++ aws_lambda_powertools/event_handler/content_types.py
@@ -1,0 +1,2 @@
+APPLICATION_JSON = "application/json"
+PLAIN_TEXT = "plain/text"
